refactor(test helpers): Remove test helpers from exports and build PE-701
